Change syntax + begin acpi

+++ OS/kernel/src/lib/print.h
@@ -2,4 +2,4 @@
 
 #include <types.h>
 
-uint32_t vprintf(char* buffer, uint32_t bufferSize, const char* fmt, va_list args);
+size_t vprintf(char* buffer, size_t bufferSize, const char* fmt, va_list args);
